chore: refactor catalog management by integrating shared handlers for add, remove, and reorder actions, enhancing user experience and code maintainability
